requirements.txt
numpy==1.23.1
pandas==1.4.3
scikit_learn==1.1.1
streamlit==1.11.1
tensorflow-cpu==2.9.1
